spr's scale parameter is now a float instead of an int, so spr(1, x, y, scale: 0.5f) shrinks and spr(1, x, y, scale: 4f) enlarges. The build succeeds (only pre-existing nullable-annotation warnings).
The change touches three places:

IMono8API.cs:47 — interface signature, float scale = 1f
Mono8API.cs:194 — the spr implementation, which just forwards the value
SpriteSheet.cs:557-576 — where the destination rectangle is actually computed
In SpriteSheet.Draw I clamp scale to [0.125f, 8f] (exposed as MinScale/MaxScale constants) so the 1/8× through 8× range you described is what actually gets honored, and anything wilder gets pinned to the edge rather than producing a degenerate rectangle. The destination width and height are rounded rather than truncated, and floored at 1 pixel — at scale: 0.125f an 8×8 tile lands exactly on 1×1, but a rounding wobble on a non-power-of-two scale could otherwise collapse a dimension to zero and draw nothing.

Existing callers that pass an int (for example the reference-sprite draw at SpriteEditor.cs:908) keep working through implicit int-to-float widening, so nothing else needed updating.

Two things worth flagging. First, this is a source-compatible but binary-breaking change to a public interface — anything compiled against the old int scale overload needs a rebuild. Second, I only verified this compiles; I didn't launch the app to look at a downscaled sprite. Point sampling at fractional scales tends to drop pixel rows unevenly (a 0.5× 8×8 tile keeps 4 of the 8 rows, and which 4 depends on the sampler), so if downscaling is meant to look clean rather than just small, that's worth eyeballing before relying on it.

+++ src/IMono8API.cs
@@ -44,7 +44,7 @@ public interface IMono8API
         int y,
         int width = 1,
         int height = 1,
-        int scale = 1,
+        float scale = 1f,
         bool flipX = false,
         bool flipY = false,
         float colorOpaqueness = 1f);
